LoadConfig: Parse configuration file espeak-ng-data/config for language options.
This removes the ability to parse for "soundicon". It will be readded
later.

Using this option requires a file espeak-ng-data/config. It is read
exactly like a language file. It overwrites the existing config.

Notes:
Reading and parsing the file is duplicate code with LoadVoice() and
should be refactored.

This feature has no tests yet.