prevent data loss when converting intptr_t to int
q[2] and q[3] are of type intptr_t, i.e. typically 64-bit, but MarkerEvent expects 32-bit ints instead, so that the 4 upper bytes are lost in the conversion. In the case of WCMD_MARKER, these four discarded bytes could contain phoneme codes. This case is somewhat rare on little-endian systems, where "upper" means 'last', but usual on big-endian systems, where "upper" means 'first' (we noticed the bug when testing on a s390x processor). The fix works on 64-bit (where all 8 phoneme bytes are in q[2]) as well as on 32-bit (where the 8 phoneme bytes are distributed over q[2] and q[3]).